+++ WinQuake/mathlib.h
@@ -23,9 +23,9 @@ typedef float vec_t;
 typedef vec_t vec3_t[3];
 typedef vec_t vec5_t[5];
 
-typedef int fixed4_t;
-typedef int fixed8_t;
-typedef int fixed16_t;
+typedef long fixed4_t;
+typedef long fixed8_t;
+typedef long fixed16_t;
 
 #ifndef M_PI
 #define M_PI 3.14159265358979323846 // matches value in gcc v2 math.h
