+++ WinQuake/r_misc.c
@@ -395,8 +395,8 @@ r_refdef.viewangles[2]=    0;
 
   if ((r_dowarp != r_dowarpold) || r_viewchanged || lcd_x.value) {
     if (r_dowarp) {
-      if ((vid.width <= vid.maxwarpwidth) &&
-          (vid.height <= vid.maxwarpheight)) {
+      if (((int) vid.width <= vid.maxwarpwidth) &&
+          ((int) vid.height <= vid.maxwarpheight)) {
         vrect.x = 0;
         vrect.y = 0;
         vrect.width = vid.width;
